refactor(desktop): reorganize harness into per-provider packages
Move opencode and pi adapters out of harness/adapters/ into sibling
provider folders, document why harness/ sits beside api/ in hexagonal
layout, and add hermes/ stub with factory registration guidance.
