--- libs/naas-abi/naas_abi/apps/desktop/harness/adapters/__init__.py
@@ -1,1 +1,0 @@
-"""Harness adapters (secondary): opencode, pi."""
