+++ libs/naas-abi/naas_abi/apps/desktop/harness/hermes/__init__.py
@@ -1,0 +1,1 @@
+"""Hermes harness adapter (not implemented)."""
