+++ libs/naas-abi/naas_abi/apps/desktop/harness/opencode/__init__.py
@@ -1,0 +1,5 @@
+"""Opencode harness adapter."""
+
+from .adapter import OpencodeHarnessAdapter
+
+__all__ = ["OpencodeHarnessAdapter"]
