+++ libs/naas-abi/naas_abi/apps/desktop/harness/opencode/adapter.py
