+++ libs/naas-abi/naas_abi/apps/desktop/harness/pi/__init__.py
@@ -1,0 +1,5 @@
+"""pi harness adapter."""
+
+from .adapter import PiHarnessAdapter
+
+__all__ = ["PiHarnessAdapter"]
