+++ libs/naas-abi/naas_abi/apps/desktop/harness/pi/adapter.py
